+++ index.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 "use strict";
 
 const Prism = require("prismjs");
@@ -6,7 +8,15 @@ const loadLanguages = require("prismjs/components/index.js");
 // Load every available Prism.js language.
 loadLanguages();
 
-module.exports = function babelPluginSyntaxHighlight({ types }) {
+/**
+ * A [Babel](https://babeljs.io) plugin that transforms the code contents of
+ * template literals lead by comments specifying a
+ * [Prism.js](https://prismjs.com) language into syntax highlighted HTML.
+ * @param {import("@babel/core") & import("@babel/core").ConfigAPI} babelApi
+ *   Babel API.
+ * @returns {import("@babel/core").PluginObj} Babel plugin.
+ */
+function babelPluginSyntaxHighlight({ types }) {
   return {
     name: "syntax-highlight",
     visitor: {
@@ -34,6 +44,13 @@ module.exports = function babelPluginSyntaxHighlight({ types }) {
         // Skip this template literal if there are no leading comments.
         if (!leadingComments) return;
 
+        /**
+         * @type {Array<{
+         *   leadingCommentIndex: number,
+         *   comment: import("@babel/core").types.Comment,
+         *   languageName: string
+         * }>}
+         */
         const languageSpecifierComments = [];
 
         leadingComments.forEach((comment, leadingCommentIndex) => {
@@ -51,6 +68,7 @@ module.exports = function babelPluginSyntaxHighlight({ types }) {
 
           if (match) {
             if (!match[1])
+              // @ts-ignore Incorrect types.
               throw state.buildCodeFrameError(
                 comment,
                 "Missing the Prism.js language name."
@@ -66,21 +84,36 @@ module.exports = function babelPluginSyntaxHighlight({ types }) {
 
         if (languageSpecifierComments.length) {
           if (languageSpecifierComments.length > 1)
+            // @ts-ignore Incorrect types.
             throw state.buildCodeFrameError(
               languageSpecifierComments[1].comment,
               "Multiple Prism.js language names specified."
             );
 
-          const prismLanguage =
-            Prism.languages[languageSpecifierComments[0].languageName];
+          const { languageName } = languageSpecifierComments[0];
 
-          if (!prismLanguage)
+          if (
+            !(languageName in Prism.languages) ||
+            // Guard against the edge case where the specified language name is
+            // the name of one of the `Prism.languages` methods, e.g. `extend`.
+            typeof Prism.languages[
+              /** @type {keyof typeof Prism.languages} */ (languageName)
+            ] !== "object"
+          )
+            // @ts-ignore Incorrect types.
             throw state.buildCodeFrameError(
               languageSpecifierComments[0].comment,
-              `\`${languageSpecifierComments[0].languageName}\` isn’t an available Prism.js language name.`
+              `\`${languageName}\` isn’t an available Prism.js language name.`
             );
 
+          const prismLanguage = /** @type {import("prismjs").Grammar} */ (
+            Prism.languages[
+              /** @type {keyof typeof Prism.languages} */ (languageName)
+            ]
+          );
+
           if (path.node.expressions.length)
+            // @ts-ignore Incorrect types.
             throw state.buildCodeFrameError(
               path.node.expressions[0],
               "Template literal expressions aren’t supported."
@@ -89,8 +122,10 @@ module.exports = function babelPluginSyntaxHighlight({ types }) {
           // Create the syntax highlighted code string.
           const highlightedCode = Prism.highlight(
             // Get the code string from the template literal.
+            // @ts-ignore
             path.node.quasis[0].value.cooked,
-            prismLanguage
+            prismLanguage,
+            languageName
           );
 
           // Remove the language specifier comment. This has to happen before
@@ -123,4 +158,6 @@ module.exports = function babelPluginSyntaxHighlight({ types }) {
       },
     },
   };
-};
+}
+
+module.exports = babelPluginSyntaxHighlight;