+++ changelog.md
@@ -1,5 +1,15 @@
 # babel-plugin-syntax-highlight changelog
 
+## Next
+
+### Major
+
+- Updated Node.js support to `^12.22.0 || ^14.17.0 || >= 16.0.0`.
+
+### Patch
+
+- Also run GitHub Actions CI with Node.js v17, and drop v15.
+
 ## 3.0.0
 
 ### Major