+++ changelog.md
@@ -5,9 +5,11 @@
 ### Major
 
 - Updated Node.js support to `^12.22.0 || ^14.17.0 || >= 16.0.0`.
+- Updated dev dependencies, some of which require newer Node.js versions than previously supported.
 
 ### Patch
 
+- Updated dependencies.
 - Also run GitHub Actions CI with Node.js v17, and drop v15.
 
 ## 3.0.0