+++ changelog.md
@@ -14,6 +14,7 @@
 - Also run GitHub Actions CI with Node.js v17, and drop v15.
 - Simplified package scripts.
 - Configured Prettier option `singleQuote` to the default, `false`.
+- Improved documentation.
 
 ## 3.0.0