+++ changelog.md
@@ -15,6 +15,8 @@
 - Simplified package scripts.
 - Updated GitHub Actions CI config:
   - Run tests with Node.js v14, v16, v18.
+  - Updated `actions/checkout` to v3.
+  - Updated `actions/setup-node` to v3.
 - Configured Prettier option `singleQuote` to the default, `false`.
 - Improved documentation.